+++ koin-projects/koin-android-viewmodel/src/main/java/org/koin/android/viewmodel/ext/android/LifecycleOwnerExt.kt
@@ -32,15 +32,27 @@ import kotlin.reflect.KClass
  */
 
 /**
+ * Lazy get a viewModel instance
+ *
+ * @param qualifier - Koin BeanDefinition qualifier (if have several ViewModel beanDefinition of the same type)
+ * @param parameters - parameters to pass to the BeanDefinition
+ * @param clazz
+ */
+fun <T : ViewModel> LifecycleOwner.viewModel(
+        clazz: KClass<T>,
+        qualifier: Qualifier? = null,
+        parameters: ParametersDefinition? = null
+): Lazy<T> = lazy { getViewModel(clazz, qualifier, parameters) }
+
+/**
  * Lazy getByClass a viewModel instance
  *
  * @param qualifier - Koin BeanDefinition qualifier (if have several ViewModel beanDefinition of the same type)
- * @param scope - used scope Instance
  * @param parameters - parameters to pass to the BeanDefinition
  */
 inline fun <reified T : ViewModel> LifecycleOwner.viewModel(
-    qualifier: Qualifier? = null,
-    noinline parameters: ParametersDefinition? = null
+        qualifier: Qualifier? = null,
+        noinline parameters: ParametersDefinition? = null
 ): Lazy<T> = lazy { getViewModel<T>(qualifier, parameters) }
 
 /**
@@ -50,17 +62,10 @@ inline fun <reified T : ViewModel> LifecycleOwner.viewModel(
  * @param parameters - parameters to pass to the BeanDefinition
  */
 inline fun <reified T : ViewModel> LifecycleOwner.getViewModel(
-    qualifier: Qualifier? = null,
-    noinline parameters: ParametersDefinition? = null
+        qualifier: Qualifier? = null,
+        noinline parameters: ParametersDefinition? = null
 ): T {
-    return (this as ComponentCallbacks).getKoin().getViewModel(
-            ViewModelParameters(
-                    T::class,
-                    this@getViewModel,
-                    qualifier,
-                    parameters = parameters
-            )
-    )
+    return getViewModel(T::class, qualifier, parameters)
 }
 
 private fun LifecycleOwner.getKoin() = (this as ComponentCallbacks).getKoin()
@@ -70,16 +75,14 @@ private fun LifecycleOwner.getKoin() = (this as ComponentCallbacks).getKoin()
  *
  * @param clazz - Class of the BeanDefinition to retrieve
  * @param qualifier - Koin BeanDefinition qualifier (if have several ViewModel beanDefinition of the same type)
- * @param scope - used scope Instance
  * @param parameters - parameters to pass to the BeanDefinition
  */
 fun <T : ViewModel> LifecycleOwner.getViewModel(
-    clazz: KClass<T>,
-    qualifier: Qualifier? = null,
-    parameters: ParametersDefinition? = null
+        clazz: KClass<T>,
+        qualifier: Qualifier? = null,
+        parameters: ParametersDefinition? = null
 ): T {
-    val koin = getKoin()
-    return koin.getViewModel(
+    return getKoin().getViewModel(
             ViewModelParameters(
                     clazz,
                     this@getViewModel,